refactor(consumer): add utils
